Move DeferredNode to separate file
